Food, hunger, and coins
Given the user 5x food to start and made it so if the fish is pressed the cat is fed. If the cat is full then the user can't feed the cat any more. Also displayed the user's coins to start, though I don't think my method is going to work so it will have to be changed.